+++ CHANGELOG.md
@@ -40,6 +40,8 @@
       `opsFromAText()` generator function.
     * `newOp()`: Deprecated in favor of the new `Op` class.
     * `opAssembler()`: Deprecated in favor of the new `serializeOps()` function.
+    * `mergingOpAssembler()`: Deprecated in favor of the new `squashOps()`
+      generator function (combined with `serializeOps()`).
     * `opIterator()`: Deprecated in favor of the new `deserializeOps()`
       generator function. Also, the unused start index parameter has been
       removed, as has the unused `lastIndex()` method on the returned object.